+++ tests/testthat/test-misc.R
@@ -30,3 +30,18 @@ test_that("diff method works properly", {
   for (i in 1:(length(y)-1)) y[i] <- y[i+1] - y[i]
   expect_equal(diff(x), y[1:(length(x)-1)])
 })
+
+test_that("matrix methods work properly", {
+  x <- set_errors(1:6, 1:6)
+  xm <- as.matrix(x)
+
+  expect_equal(dim(xm), c(6, 1))
+  expect_equal(as.numeric(xm), as.numeric(x))
+  expect_equal(errors(xm), errors(x))
+
+  xt <- set_errors(c(1, 4, 2, 5, 3, 6), c(1, 4, 2, 5, 3, 6))
+  xm <- set_errors(cbind(1:3, 4:6), 1:6)
+
+  expect_equal(as.numeric(xt), as.numeric(t(xm)))
+  expect_equal(errors(xt), errors(t(xm)))
+})