+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Minor changes and fixes:
 
+* `Inf`, `NaN` and `NA` values default to `Inf`, `NaN` and `NA` errors respectively (#10).
 * Fix errors defined as integers (#12).
 * Do not allow boolean operators on `errors` objects (#13).
 * Give a default value to `set_errors` (#14).